[cmake] Properly export plugin data
