+++ .gitignore
@@ -1,0 +1,4 @@
+build/
+.ccls-cache/
+.cache/
+compile_commands.json
