+++ .gitmodules
@@ -1,0 +1,3 @@
+[submodule "cmake"]
+	path = cmake
+	url = git@github.com:jrl-umi3218/jrl-cmakemodules
