+++ include/mc_neuron_mocap_plugin/MoCapProperties.h
@@ -1,0 +1,73 @@
+#pragma once
+
+enum MoCap_Body_part
+{
+  Hips,
+  RightUpLeg,
+  RightLeg,
+  RightFoot,
+  LeftUpLeg,
+  LeftLeg,
+  LeftFoot,
+  RightShoulder,
+  RightArm,
+  RightForeArm,
+  RightHand,
+  LeftShoulder,
+  LeftArm,
+  LeftForeArm,
+  LeftHand,
+  Head,
+  Neck1,
+  Neck,
+  Spine2,
+  Spine1,
+  Spine,
+  RightHandThumb1,
+  RightHandThumb2,
+  RightHandThumb3,
+  RightInHandIndex,
+  RightHandIndex1,
+  RightHandIndex2,
+  RightHandIndex3,
+  RightInHandMiddle,
+  RightHandMiddle1,
+  RightHandMiddle2,
+  RightHandMiddle3,
+  RightInHandRing,
+  RightHandRing1,
+  RightHandRing2,
+  RightHandRing3,
+  RightInHandPinky,
+  RightHandPinky1,
+  RightHandPinky2,
+  RightHandPinky3,
+  LeftHandThumb1,
+  LeftHandThumb2,
+  LeftHandThumb3,
+  LeftInHandIndex,
+  LeftHandIndex1,
+  LeftHandIndex2,
+  LeftHandIndex3,
+  LeftInHandMiddle,
+  LeftHandMiddle1,
+  LeftHandMiddle2,
+  LeftHandMiddle3,
+  LeftInHandRing,
+  LeftHandRing1,
+  LeftHandRing2,
+  LeftHandRing3,
+  LeftInHandPinky,
+  LeftHandPinky1,
+  LeftHandPinky2,
+  LeftHandPinky3
+};
+
+enum MoCap_Parameters
+{
+  MoCap_Position,
+  MoCap_Velocity,
+  MoCap_Quaternion,
+  MoCap_Accelerated_Velocity,
+  MoCap_Gyro
+};
